fix(feature): wire up label picker on feature detail page
The Labels row in the feature Properties panel rendered as static text, so
features could not be tagged from the UI. Extract the ticket page's inline
label picker into a shared LabelsCombobox component and use it on both the
feature and ticket detail pages, backed by the existing tag.setFeatureTags
mutation.
