+++ src/app/(sidemenu)/w/[workspaceSlug]/products/[productSlug]/tickets/[ticketId]/page.tsx
@@ -53,13 +53,13 @@ import {
 import { generateLinearId } from "~/lib/fun-ids";
 import { PriorityIcon } from "~/app/_components/product/PriorityIcon";
 import { TicketDependenciesSection } from "~/app/_components/product/TicketDependenciesSection";
+import { LabelsCombobox } from "~/app/_components/product/LabelsCombobox";
 import {
   STATUS_OPTIONS,
   STATUS_COLORS,
   type TicketStatus,
 } from "~/lib/ticket-statuses";
 import { TagBadge } from "~/app/_components/TagBadge";
-import { getTagMantineColor } from "~/utils/tagColors";
 import { MarkdownRenderer } from "~/app/_components/shared/MarkdownRenderer";
 
 // ---------------------------------------------------------------------------
@@ -194,98 +194,6 @@ function EpicCombobox({
           {!search.trim() && filtered.length === 0 && (
             <Combobox.Empty>
               <Text size="xs" className="text-text-muted">No epics</Text>
-            </Combobox.Empty>
-          )}
-        </Combobox.Options>
-      </Combobox.Dropdown>
-    </Combobox>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Combobox: Labels (multi-select with create)
-// ---------------------------------------------------------------------------
-
-function LabelsCombobox({
-  selectedIds,
-  allTags,
-  ticketTags,
-  onChange,
-  onCreate,
-}: {
-  selectedIds: string[];
-  allTags: Array<{ id: string; name: string; color: string }>;
-  ticketTags: Array<{ tag: { id: string; name: string; color: string } }>;
-  onChange: (tagIds: string[]) => void;
-  onCreate: (name: string) => void;
-}) {
-  const combobox = useCombobox({ onDropdownClose: () => { combobox.resetSelectedOption(); setSearch(""); } });
-  const [search, setSearch] = useState("");
-
-  const filtered = allTags.filter((t) => t.name.toLowerCase().includes(search.toLowerCase().trim()));
-  const exactMatch = allTags.some((t) => t.name.toLowerCase() === search.toLowerCase().trim());
-
-  const toggleTag = (tagId: string) => {
-    if (selectedIds.includes(tagId)) {
-      onChange(selectedIds.filter((id) => id !== tagId));
-    } else {
-      onChange([...selectedIds, tagId]);
-    }
-  };
-
-  return (
-    <Combobox store={combobox} onOptionSubmit={(val) => {
-      if (val === "__create") {
-        onCreate(search.trim());
-        combobox.closeDropdown();
-      } else {
-        toggleTag(val);
-        // Keep dropdown open for multi-select
-      }
-    }}>
-      <Combobox.Target>
-        <div
-          className="cursor-pointer min-h-[24px] flex items-center"
-          onClick={() => combobox.toggleDropdown()}
-        >
-          {ticketTags.length > 0 ? (
-            <Group gap={4}>
-              {ticketTags.map((t) => (
-                <TagBadge key={t.tag.id} tag={t.tag} size="xs" />
-              ))}
-            </Group>
-          ) : (
-            <Text size="xs" className="text-text-muted">None</Text>
-          )}
-        </div>
-      </Combobox.Target>
-      <Combobox.Dropdown>
-        <Combobox.Search
-          value={search}
-          onChange={(e) => { setSearch(e.currentTarget.value); combobox.updateSelectedOptionIndex(); }}
-          placeholder="Search or create..."
-          size="xs"
-        />
-        <Combobox.Options>
-          {filtered.map((tag) => {
-            const isSelected = selectedIds.includes(tag.id);
-            return (
-              <Combobox.Option key={tag.id} value={tag.id} active={isSelected}>
-                <div className="flex items-center gap-2">
-                  {isSelected && <CheckIcon size={12} />}
-                  <Badge size="xs" variant="light" color={getTagMantineColor(tag.color)}>{tag.name}</Badge>
-                </div>
-              </Combobox.Option>
-            );
-          })}
-          {search.trim() && !exactMatch && (
-            <Combobox.Option value="__create">
-              <Text size="xs" className="text-blue-400">+ Create &quot;{search.trim()}&quot;</Text>
-            </Combobox.Option>
-          )}
-          {!search.trim() && filtered.length === 0 && (
-            <Combobox.Empty>
-              <Text size="xs" className="text-text-muted">No labels</Text>
             </Combobox.Empty>
           )}
         </Combobox.Options>
@@ -758,7 +666,7 @@ export default function TicketDetailPage() {
           <LabelsCombobox
             selectedIds={ticket.tags?.map((t: { tag: { id: string } }) => t.tag.id) ?? []}
             allTags={tags?.allTags ?? []}
-            ticketTags={ticket.tags ?? []}
+            entityTags={ticket.tags ?? []}
             onChange={(tagIds) => setTicketTags.mutate({ ticketId, tagIds })}
             onCreate={(name) => {
               if (workspaceId) createTag.mutate({ name, color: "avatar-blue", workspaceId });
